% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnbase-extensions.R
\docType{methods}
\name{correctIntensity,MSnSet-method}
\alias{correctIntensity}
\alias{correctIntensity,MSnSet-method}
\title{Correct Intensity}
\usage{
\S4method{correctIntensity}{MSnSet}(object, \ldots)
}
\arguments{
\item{object}{An \code{\linkS4class{MSnSet}} object.}

\item{\ldots}{further arguments, not used yet.}
}
\value{
A \code{\linkS4class{MSnSet}} with corrected intensity values
(\code{exprs}).
}
\description{
This method corrects the intensity values of an \code{\linkS4class{MSnSet}}
object by applying the intensity model built in the synapter workflow by
\code{\link{modelIntensity}}.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\seealso{
MSnSet documentation: \code{\linkS4class{MSnSet}}
}

