\name{Deprecated}
\title{synapter Deprecated and Defunct}

\alias{synapter-deprecated}
\alias{synapter-defunct}

\alias{synapterGUI}

\description{
  
  The function, class, or data object you have asked for have been
  deprecated or made defunct. Deprecated functions are provided for
  compatibility with older versions of the \code{synapter} package only,
  and will be defunct at the next release. Defunct functions have been
  removed. 

  Deprecated: 
  
  Defunct: \code{synapterGUI}; use the replacement \code{\link{synergise}}  
  
}

\keyword{internal}
