% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requantify-functions.R
\docType{methods}
\name{rescaleForTop3,MSnSet,MSnSet-method}
\alias{rescaleForTop3,MSnSet,MSnSet-method}
\alias{rescaleForTop3}
\title{Rescale for TOP3}
\usage{
\S4method{rescaleForTop3}{MSnSet,MSnSet}(before, after, saturationThreshold,
onlyForSaturatedRuns=TRUE, \ldots)
}
\arguments{
\item{before}{An \code{\linkS4class{MSnSet}} object before requantification.}

\item{after}{The same \code{\linkS4class{MSnSet}} object as \code{before} but
after requantification.}

\item{saturationThreshold}{\code{double}, intensity of an ion (isotope of a
given charge state) at which saturation is starting to occur.}

\item{onlyForSaturatedRuns}{\code{logical}, rescale just runs where at least
one isotope is affected by saturation.}

\item{\ldots}{further arguments passed to internal functions. Currently
ignored.}
}
\value{
\code{\linkS4class{MSnSet}} where the
\code{assayData} are requantified.
}
\description{
This method rescales the intensity values of an \code{\linkS4class{MSnSet}}
object to be suiteable for TOP3 quantification.
}
\details{
If an \code{\linkS4class{MSnSet}} object was requantified using the
\code{method="sum"} requantification method
(see \code{\link{requantify,MSnSet-method}}) TOP3 is not valid anymore
because the most abundant proteins are penalised by removing high intensity
isotopes.

To overcome this \code{rescaleForTop3} takes the proportion
\code{isotope/sum(isotopes} for each requantified peptide and calculates a
correction factor by comparing these proportions against the unsaturated
isotopes before requantification. The new rescale intensity values of the
isotopes are the mean correction factor multiplied with the corrected
intensity values (see
\url{https://github.com/lgatto/synapter/issues/39#issuecomment-207987278} for
the complete explanation/discussion).
}
\references{
See discussion on github: \url{https://github.com/lgatto/synapter/issues/39}
}
\seealso{
MSnSet documentation: \code{\linkS4class{MSnSet}}
\url{https://github.com/lgatto/synapter/issues/39#issuecomment-207987278}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de} and Pavel Shliaha
}
