% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_tools.R
\name{synapterGuide}
\alias{synapterGuide}
\title{Opens the 'synapter' vignette}
\usage{
synapterGuide()
}
\description{
Opens the relevant vignette; a shortcut to using \code{vignette}.
\code{synapterGuide()} gives access to the main overview vignette.
}
\author{
Laurent Gatto
}

