% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnbase-extensions.R
\docType{methods}
\name{synapterPlgsAgreement,MSnSet-method}
\alias{synapterPlgsAgreement,MSnSet-method}
\alias{synapterPlgsAgreement}
\title{Synapter/PLGS Agreement}
\usage{
\S4method{synapterPlgsAgreement}{MSnSet}(object, \ldots)
}
\arguments{
\item{object}{An \code{\linkS4class{MSnSet}} object.}

\item{\ldots}{further arguments, not used yet.}
}
\value{
\code{\linkS4class{MSnSet}} where the columns \code{nIdentified},
\code{nAgree}, \code{nDisagree} and \code{synapterPlgsAgreementRatio} were
added to the feature data.
}
\description{
This method checks the agreement between synapter analysis and PLGS results.
}
\details{
Each synapter object has \code{synapterPlgsAgreement} column in its
\code{MatchedEMRTs} \code{data.frame} (see \code{\link{writeMatchedEMRTs}}).
After converting the synapter object into an \code{\linkS4class{MSnSet}}
instance via \code{as(synapterobject, "MSnSet"} this column could be find in
the feature data (\code{fData(msnset)$synapterPlgsAgreement}).\cr
In the \code{synapterPlgsAgreement} each peptide is classified as:
\itemize{
 \item{\code{"agree"}: EMRT identified in identification and quantitation run
 by PLGS and same EMRT matched in synapter's grid search.}
 \item{\code{"disagree"}: EMRT identified in identification and quantitation
 run by PLGS and a different EMRT was matched in synapter's grid search.}
 \item{\code{"no_plgs_id"}: EMRT was \emph{not} identified in the
 quantitation run by PLGS but matched in synapter's grid search.}
 \item{\code{"no_synapter_transfer"}: EMRT was identified in the
 identification and quantitation run by PLGS but not matched in synapter's
 grid search.}
 \item{\code{"no_id_or_transfer"}: EMRT was \emph{not} identified in the
 quantitation run by PLGS and not matched in synapter's grid search.}
 \item{\code{"multiple_ident_matches"}: a single quantitation EMRT was
 matched by synapter to multiple identification EMRTs found by PLGS (could
 happen if the grid search parameters are too relaxed).}
}
After combining multiple \code{\linkS4class{MSnSet}} the method
\code{synapterPlgsAgreement} adds additional columns to the feature data:
\itemize{
 \item{\code{nIdentified}: how often a peptide was identified across
 multiple runs?}
 \item{\code{nAgree}: how often a peptide was identified by PLGS and synapter
 across multiple runs (counts \code{"agree"} entries)?}
 \item{\code{nDisagree}: how often a peptide was differently identified by
 PLGS and synapter across multiple runs (counts \code{"disagree"} entries)?}
 \item{\code{synapterPlgsAgreementRatio}: \code{nAgree/(nAgree +
 nDisagree)}.}
}
}
\references{
See discussion on github: \url{https://github.com/lgatto/synapter/issues/73}
}
\seealso{
MSnSet documentation: \code{\linkS4class{MSnSet}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
