% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synapterGUI}
\alias{synapterGUI}
\title{Launch the synapter GUI}
\usage{
synapterGUI(n = 1)
}
\arguments{
\item{n}{Number of analysis to be run.}
}
\value{
Invisibly returns \code{null}. Used for its side effects.
}
\description{
Note: The tcltk interface is being deprecated. Please consider
using its underlying \code{synergise()} function instead.
}
\details{
This function starts the synapter graphical user interface
that allows to (1) load \code{n} sets of files to be analyses,
(2) parameters used for data filtering and retention time
modelling and (3) parameters used for grid search.
The \code{n} file sets and the parameters recorded by the GUI
are used to call \code{\link{synergise}} that performs the
analysis and generates \code{n} html reports.
See the \code{synapter} vignette for a description of the
interface and \code{\link{synergize}} for a description of the
parameters and data processing.

The GUI code requires the Tcl package 'BWidget' to display the tree
drill-down file selection menu. The package is available on Windows
(in case of issues, please read
\url{http://www.stat.berkeley.edu/users/spector/s133/Bwidget.html}).
On GNU/Linux, the package needs to be installed separately;
on Ubuntu, the package is called 'bwidget'.
}
\author{
Laurent Gatto
}

