% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synapterGuide}
\alias{synapterGuide}
\title{Opens the 'synapter' vignette}
\usage{
synapterGuide()
}
\description{
Opens the relevant vignette; a shortcut to using \code{vignette}.
\code{synapterGuide()} gives access to the main overview vignette.
}
\author{
Laurent Gatto
}

